# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenNotebook",
    "description": "Collections of sources, and the decks, audio overviews, mind maps and notes made from them. Everything the web app does goes through this API.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/collections/{cid}/sources": {
      "get": {
        "tags": [
          "sources"
        ],
        "summary": "List Sources",
        "description": "The sources in a collection, in the order they were added, a page at\na time: `X-Next-Offset` says where the next page starts when there is\none.",
        "operationId": "list_sources",
        "parameters": [
          {
            "name": "cid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Cid"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "description": "The most sources in one answer; 500 unless set",
              "default": 500,
              "title": "Limit"
            },
            "description": "The most sources in one answer; 500 unless set"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100000,
              "minimum": 0,
              "description": "Where this page starts: the `X-Next-Offset` header of the page before",
              "default": 0,
              "title": "Offset"
            },
            "description": "Where this page starts: the `X-Next-Offset` header of the page before"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SourceOut"
                  },
                  "title": "Response List Sources"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "sources"
        ],
        "summary": "Add Sources",
        "description": "Add a note, or read web pages, into a collection. One result per thing\nasked for, failures included, each saying why.",
        "operationId": "add_sources",
        "parameters": [
          {
            "name": "cid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Cid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/AddNote"
                  },
                  {
                    "$ref": "#/components/schemas/AddUrls"
                  }
                ],
                "discriminator": {
                  "propertyName": "kind",
                  "mapping": {
                    "text": "#/components/schemas/AddNote",
                    "urls": "#/components/schemas/AddUrls"
                  }
                },
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AddResult"
                  },
                  "title": "Response Add Sources"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/collections/{cid}/sources/files": {
      "post": {
        "tags": [
          "sources"
        ],
        "summary": "Add Files",
        "description": "Add documents to a collection, at most ten at a time. Each is read\ninto text verbatim and titled by its file name; a result says why when\none could not be read.",
        "operationId": "add_files",
        "parameters": [
          {
            "name": "cid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Cid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_add_files"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AddResult"
                  },
                  "title": "Response Add Files"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/collections/{cid}/sources/{name}": {
      "get": {
        "tags": [
          "sources"
        ],
        "summary": "Read Source",
        "description": "One source with its text, for reading it beside a citation.",
        "operationId": "read_source",
        "parameters": [
          {
            "name": "cid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Cid"
            }
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceText"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "sources"
        ],
        "summary": "Remove Source",
        "description": "Remove one source from a collection.",
        "operationId": "remove_source",
        "parameters": [
          {
            "name": "cid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Cid"
            }
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/search": {
      "post": {
        "tags": [
          "sources"
        ],
        "summary": "Web Search",
        "description": "Search the web. Nothing is added: pass the best links to add_sources.\nA few seconds; charged to the person like any model call.",
        "operationId": "web_search",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WebHit"
                  },
                  "type": "array",
                  "title": "Response Web Search"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/collections/{cid}/research": {
      "post": {
        "tags": [
          "sources"
        ],
        "summary": "Deep Research",
        "description": "Research a topic in depth and add the written report as one source.\nRuns in the background, minutes; follow it on /api/jobs/{job_id}.",
        "operationId": "deep_research",
        "parameters": [
          {
            "name": "cid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Cid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResearchReq"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/shares/{share_id}/sources/{name}": {
      "get": {
        "tags": [
          "shares"
        ],
        "summary": "Read Shared Source",
        "description": "One source of a share that includes its sources, with its text.",
        "operationId": "read_shared_source",
        "parameters": [
          {
            "name": "share_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Share Id"
            }
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceText"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddNote": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "text",
            "title": "Kind"
          },
          "text": {
            "type": "string",
            "maxLength": 2000000,
            "title": "Text",
            "description": "The note, in plain text or Markdown"
          },
          "title": {
            "type": "string",
            "maxLength": 200,
            "title": "Title",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "kind",
          "text"
        ],
        "title": "AddNote"
      },
      "AddResult": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url",
            "description": "The page asked for; empty for a note or a file"
          },
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "source": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SourceOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "type": "string",
            "title": "Error",
            "description": "Why it was not added, in a sentence; empty when ok"
          }
        },
        "type": "object",
        "required": [
          "url",
          "ok",
          "error"
        ],
        "title": "AddResult",
        "description": "What happened to one thing asked to be added."
      },
      "AddUrls": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "urls",
            "title": "Kind"
          },
          "urls": {
            "items": {
              "type": "string",
              "maxLength": 2083,
              "minLength": 1,
              "format": "uri"
            },
            "type": "array",
            "maxItems": 8,
            "minItems": 1,
            "title": "Urls"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "urls"
        ],
        "title": "AddUrls"
      },
...
      "Body_add_files": {
        "properties": {
          "files": {
            "items": {
              "type": "string",
              "contentMediaType": "application/octet-stream"
            },
            "type": "array",
            "title": "Files",
            "description": "pdf, docx, pptx, xlsx, md, txt or csv"
          }
        },
        "type": "object",
        "required": [
          "files"
        ],
        "title": "Body_add_files"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "done",
              "failed",
              "cancelled"
            ],
            "title": "Status"
          },
          "step": {
            "type": "string",
            "title": "Step",
            "description": "What it is doing now; empty before it starts"
          },
          "steps_done": {
            "type": "integer",
            "title": "Steps Done"
          },
          "steps_total": {
            "type": "integer",
            "title": "Steps Total",
            "description": "0 when the work does not report its steps"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error",
            "description": "Why it failed or stopped; null unless it did"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "collection_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Collection Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "waiting": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Waiting",
            "description": "Why queued work has not started, in a sentence: no worker is running. Null otherwise"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "status",
          "step",
          "steps_done",
          "steps_total",
          "error",
          "session_id",
          "collection_id",
          "created_at"
        ],
        "title": "JobOut",
        "description": "A piece of background work and how far it got."
      },
...
      "ResearchReq": {
        "properties": {
          "topic": {
            "type": "string",
            "maxLength": 400,
            "minLength": 1,
            "title": "Topic"
          }
        },
        "type": "object",
        "required": [
          "topic"
        ],
        "title": "ResearchReq"
      },
...
      "SearchReq": {
        "properties": {
          "query": {
            "type": "string",
            "maxLength": 400,
            "minLength": 1,
            "title": "Query"
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "SearchReq"
      },
...
      "SourceOut": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Stable within its collection; address the source by it"
          },
          "kind": {
            "type": "string",
            "enum": [
              "url",
              "text",
              "file",
              "research"
            ],
            "title": "Kind"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "url": {
            "type": "string",
            "title": "Url",
            "description": "Where it was read from; empty for a note, a file or a report"
          },
          "chars": {
            "type": "integer",
            "title": "Chars",
            "description": "Length of the readable text"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "words": {
            "type": "integer",
            "title": "Words",
            "description": "About how many words it holds",
            "readOnly": true
          },
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "The line under its name: where it came from and how long it is, `example.org \u00b7 1200 words` or `PDF \u00b7 300 words`",
            "readOnly": true
          },
          "icon": {
            "type": "string",
            "title": "Icon",
            "description": "The picture to lead its row with: its site's icon, loaded from the site itself; empty for a note, a file or a report",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "name",
          "kind",
          "title",
          "url",
          "chars",
          "created_at",
          "words",
          "detail",
          "icon"
        ],
        "title": "SourceOut"
      },
      "SourceText": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Stable within its collection; address the source by it"
          },
          "kind": {
            "type": "string",
            "enum": [
              "url",
              "text",
              "file",
              "research"
            ],
            "title": "Kind"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "url": {
            "type": "string",
            "title": "Url",
            "description": "Where it was read from; empty for a note, a file or a report"
          },
          "chars": {
            "type": "integer",
            "title": "Chars",
            "description": "Length of the readable text"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "text": {
            "type": "string",
            "title": "Text",
            "description": "The source verbatim, as Markdown"
          },
          "words": {
            "type": "integer",
            "title": "Words",
            "description": "About how many words it holds",
            "readOnly": true
          },
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "The line under its name: where it came from and how long it is, `example.org \u00b7 1200 words` or `PDF \u00b7 300 words`",
            "readOnly": true
          },
          "icon": {
            "type": "string",
            "title": "Icon",
            "description": "The picture to lead its row with: its site's icon, loaded from the site itself; empty for a note, a file or a report",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "name",
          "kind",
          "title",
          "url",
          "chars",
          "created_at",
          "text",
          "words",
          "detail",
          "icon"
        ],
        "title": "SourceText"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WebHit": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "snippet": {
            "type": "string",
            "title": "Snippet"
          }
        },
        "type": "object",
        "required": [
          "title",
          "url",
          "snippet"
        ],
        "title": "WebHit"
      }
    }
  }
}